pyproject.toml
[project]
name = "skills-threat-analysis"
version = "0.1.0"
description = "Prompt Injection scanner for Claude Code skills"
authors = [
    {name = "kenshin", email = "kenshin.acs@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "anthropic>=0.42.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
scan-skills = "scanner.cli:main"

[[tool.poetry.packages]]
include = "scanner"
from = "src"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
